feat: add cli options

+++ index.ts
@@ -4,18 +4,51 @@ import { Config } from './utils/utils';
 import createBaseConfig from './utils/base';
 import { createBirdConfig, createWGConfig } from './utils/bgp';
 import { createIBirdConfig, createIWGConfig } from './utils/ibgp';
+import yargs from 'yargs';
+import { hideBin } from 'yargs/helpers';
+
+const _argv = yargs(hideBin(process.argv))
+  .option('exclude-wireguard', {
+    type: 'boolean',
+    default: false,
+    description: 'Skip wireguard config generation'
+  })
+  .option('exclude-bird', {
+    type: 'boolean',
+    default: false,
+    description: 'Skip bird config generation'
+  })
+  .option('exclude-internal-config', {
+    type: 'boolean',
+    default: false,
+    description: 'Skip internal config generation'
+  })
+  .option('exclude-external-config', {
+    type: 'boolean',
+    default: false,
+    description: 'Skip external config generation'
+  })
+  .option('delete', {
+    type: 'boolean',
+    default: true,
+    description: 'Delete existing config files'
+  })
+  .parse();
 
 const outputDir = resolve(__dirname, './dist');
 
 const resolveHostDir = (path: string) => resolve(outputDir, path);
 
-const checkFileExists = (file: string) => fs
-  .access(file, constants.F_OK)
-  .then(() => true)
-  .catch(() => false);
+const checkFileExists = (file: string) =>
+  fs
+    .access(file, constants.F_OK)
+    .then(() => true)
+    .catch(() => false);
 
 (async () => {
-  if (await checkFileExists(outputDir)) {
+  const argv = _argv instanceof Promise ? await _argv : _argv;
+
+  if (argv['delete'] && (await checkFileExists(outputDir))) {
     await fs.rm(outputDir, { recursive: true });
   }
 
@@ -25,7 +58,7 @@ const checkFileExists = (file: string) => fs
   await Promise.all(
     allHosts.map(host => {
       const basePath = resolveHostDir(host);
-      return createBaseConfig(host, basePath);
+      return createBaseConfig(host, basePath, argv);
     })
   );
 
@@ -33,10 +66,18 @@ const checkFileExists = (file: string) => fs
     allHosts.map(host => {
       const basePath = resolveHostDir(host);
       return Promise.all([
-        createWGConfig(host, basePath),
-        createBirdConfig(host, basePath),
-        createIWGConfig(host, basePath),
-        createIBirdConfig(host, basePath)
+        ...(argv['excludeWireguard']
+          ? []
+          : [
+              !argv['excludeInternalConfig'] && createIWGConfig(host, basePath),
+              !argv['excludeExternalConfig'] && createWGConfig(host, basePath)
+            ]),
+        ...(argv['excludeBird']
+          ? []
+          : [
+              !argv['excludeInternalConfig'] && createIBirdConfig(host, basePath),
+              !argv['excludeExternalConfig'] && createBirdConfig(host, basePath)
+            ])
       ]);
     })
   );
